djangoapp/requirements.txt
asgiref==3.8.1
Django>=5.0.6,<5.1
psycopg2-binary>= 2.9.9,<2.10
sqlparse==0.5.0
tzdata==2024.1
djangorestframework
djangorestframework-simplejwt
Pillow
brazilcep